+++ src/section/index.js
@@ -1,4 +1,5 @@
 const Section = require('./Section');
+const parseTemplates = require('./templates');
 const find_recursive = require('../lib/recursive_match');
 
 //interpret ==heading== lines
@@ -13,7 +14,9 @@ const parse = {
 };
 const section_reg = /[\n^](={1,5}[^=]{1,200}?={1,5})/g;
 
-const parseSection = function(section, wiki, r, options) {
+const doSection = function(section, wiki, options) {
+  wiki = parseTemplates(section, wiki, options);
+
   // //parse the tables
   wiki = parse.table(section, wiki);
   // //parse the lists
@@ -32,21 +35,23 @@ const parseSection = function(section, wiki, r, options) {
   return section;
 };
 
-const makeSections = function(r, wiki, options) {
+const splitSections = function(wiki, options) {
   let split = wiki.split(section_reg); //.filter(s => s);
   let sections = [];
   for (let i = 0; i < split.length; i += 2) {
-    let title = split[i - 1] || '';
-    let txt = split[i] || '';
+    let heading = split[i - 1] || '';
+    let content = split[i] || '';
     let section = {
       title: '',
       depth: null
     };
-    section = parse.heading(section, title);
-    section = parseSection(section, txt, r, options);
+    //figure-out title/depth
+    section = parse.heading(section, heading);
+    //parse it up
+    section = doSection(section, content, options);
     sections.push(section);
   }
   return sections;
 };
 
-module.exports = makeSections;
+module.exports = splitSections;