+++ changelog.md
@@ -48,3 +48,4 @@
 * make section.title a function
 ## 3.2.0
 * add `.wikitext()` method to Document, Section, Sentence (thanks @niebert)
+* move infobox parser+data to Section class